+++ src/utils/debug.js
@@ -1,9 +1,0 @@
-function debug(message, data) {
-  if (DEBUG_MODE) {
-    if (data) {
-      console.log(message, data);
-    } else {
-      console.log(message);
-    }
-  }
-}