+++ RELEASE_NOTES.md
@@ -2,6 +2,21 @@
 
 This file contains changes based on [this template](https://github.com/palantir/plottable/wiki/Release-Notes-Template)
 
+## 0.3.0
+
+### New Features
+
+- Added a popup with the extension status for the active tab and a link to settings page
+- Internationalization of the settings page in Spanish and English
+
+### Improvements
+
+- Now tooltips are only added if there are any added node in the DOM mutation list
+
+### Bug fixes
+
+- Avoid duplication of warning tooltip DOM element
+
 ## 0.2.0
 
 ### New Features